Reverted previous commit.
All attemps to get scheduled_events will now use the cache(?)

The documenatation is not clear about the difference between
guild.scheduled_events and guild.fetch_scheduled_events.

I thought fetch_scheduled_events was the api call (and it seems to be),
but scheduled_events seems to also update when I don't expect it to.

So sometimes scheduled_events is accurate, sometimes it's not. Either way,
it's not worth forcing an api call every time I want to get the events.