Clarity commit
Variable names were easy to confuse with discord objects.
They usually hold the id that can be used to get the actual object
so I adjusted the names to reflect that.

Also, added a requirements.txt file to keep track of the dependencies.
If you look, you might be able to deduce the next feature ;)