Changed command name and fixed a display issue
Command "/prefer-map" was changed to "/prefer-maps" to reflect
its updated functionality. Previously, it was used to mark a
preference for a single map, but now it is used for all maps at once.

Fixed an issue where a map name was not being displayed in a manner
consistent with the rest of the bot's output regarding map names.