UX Changes and Dosctring adjustments
Removed the "HELP:" section of the command list. It was clunky
and not helpful. It led you to using the persistent buttons
which themselves you had to have already used to see the
help section in the command list in the first place.

Changed the style text docstring to my prefernce.
Don't quote the options and (although unimportant to the output)
ensure that it demonstrates the correct output order
of style characters in the example.