Complete rework of the event reminder system.
Event reminders only happen 1 hour before, and right on time.
Each reminder utilizes an embed for the main reminder with the 1 hour
reminder mnetioning the target role, and the on time reminder mentioning
the RSVP'ed users. Crucially, these reminders still send
"(reminder)" in the message content to ensure that the reminder
is automatically deleted by the bot after 1 day.

The new embed contains extra information that wasn't present
in the old system. It contains: An image of the map,
the number of RSVP'ed users, and the weight given to the map
from premier voting. It can contain a lot more information,
but this is the base for now.

An idea for the future is to attach a button for
viewing the map's notes, but it's fine "for now".